Update example in README (#31)
Try being more specific about how `{:system, :list, ...}` works. The current README falsely suggests that the list will be parsed into integers, which is not how it is implemented. This patch adjusts the default value in the example to the same typing as what the resolver would produce from an environment variable.